[#2131] Refactor
