+++ client/src/components/charts/bar/SimpleBarChart.jsx
@@ -5,34 +5,37 @@ import { Rect, Svg, Group, Text } from '@potion/element';
 import get from 'lodash/get';
 import { scaleLinear } from 'd3-scale';
 import { extent } from 'd3-array';
-import { AxisBottom } from '@vx/axis';
+import { AxisLeft } from '@vx/axis';
 import { Portal } from 'react-portal';
 import merge from 'lodash/merge';
-import { GridColumns } from '@vx/grid';
+import { GridRows } from '@vx/grid';
 import itsSet from 'its-set';
+import { barPropTypes, barDefaultProps } from './CommonBarChart';
 
-import { isLight } from '../../utilities/color';
-import { abbr } from '../../utilities/utils';
+import { isLight } from '../../../utilities/color';
 import {
   heuristicRound,
   replaceLabelIfValueEmpty,
   calculateMargins,
   getLabelFontSize,
   labelFitsWidth,
-} from '../../utilities/chart';
-import Legend from './Legend';
-import ResponsiveWrapper from '../common/ResponsiveWrapper';
-import ColorPicker from '../common/ColorPicker';
-import ChartLayout from './ChartLayout';
-import Tooltip from './Tooltip';
-import { labelFont, MAX_FONT_SIZE, MIN_FONT_SIZE, LABEL_CHAR_WIDTH } from '../../constants/chart';
-import RenderComplete from './RenderComplete';
+} from '../../../utilities/chart';
+import Legend from '../Legend';
+import ResponsiveWrapper from '../../common/ResponsiveWrapper';
+import ColorPicker from '../../common/ColorPicker';
+import ChartLayout from '../ChartLayout';
+import Tooltip from '../Tooltip';
+import { labelFont, MAX_FONT_SIZE, MIN_FONT_SIZE } from '../../../constants/chart';
+import RenderComplete from '../RenderComplete';
 
 const getDatum = (data, datum) => data.filter(({ key }) => key === datum)[0];
 
-const getLabelText = (label, type) => (type === 'date' ?
-  `${new Date(label * 1000)}` :
-  replaceLabelIfValueEmpty(label));
+const getLabelText = (label, type) => {
+  const o = type === 'date' ?
+    `${new Date(label * 1000)}` : replaceLabelIfValueEmpty(label);
+
+  return o;
+};
 
 const getPaddingBottom = (data, type) => {
   const labelCutoffLength = 16;
@@ -51,65 +54,9 @@ const getPaddingBottom = (data, type) => {
 
 export default class SimpleBarChart extends Component {
 
-  static propTypes = {
-    data: PropTypes.shape({
-      data: PropTypes.oneOfType([
-        PropTypes.arrayOf(
-          PropTypes.shape({
-            key: PropTypes.string,
-            value: PropTypes.number,
-          })
-        ),
-        PropTypes.arrayOf(
-          PropTypes.shape({
-            key: PropTypes.string,
-            values: PropTypes.arrayOf(
-              PropTypes.number
-            ),
-          })
-        ),
-      ]),
-      metadata: PropTypes.object,
-    }),
-    colors: PropTypes.array.isRequired,
-    colorMapping: PropTypes.object,
-    defaultColor: PropTypes.string.isRequired,
-    onChangeVisualisationSpec: PropTypes.func,
-    width: PropTypes.number.isRequired,
-    height: PropTypes.number.isRequired,
-    legendPosition: PropTypes.oneOf(['top', 'right', 'bottom', 'left', undefined]),
-    print: PropTypes.bool,
-    interactive: PropTypes.bool,
-    edit: PropTypes.bool,
-    padding: PropTypes.number,
-    marginLeft: PropTypes.number,
-    marginRight: PropTypes.number,
-    marginTop: PropTypes.number,
-    marginBottom: PropTypes.number,
-    style: PropTypes.object,
-    legendVisible: PropTypes.bool,
-    valueLabelsVisible: PropTypes.bool,
-    yAxisLabel: PropTypes.string,
-    yAxisTicks: PropTypes.number,
-    xAxisLabel: PropTypes.string,
-    grid: PropTypes.bool,
-    visualisation: PropTypes.object,
-  }
+  static propTypes = barPropTypes;
 
-  static defaultProps = {
-    interactive: true,
-    marginLeft: 180,
-    marginRight: 70,
-    marginTop: 70,
-    marginBottom: 70,
-    legendVisible: false,
-    valueLabelsVisible: false,
-    edit: false,
-    padding: 0.1,
-    colorMapping: {},
-    grid: true,
-    xAxisLabel: '',
-  }
+  static defaultProps = barDefaultProps;
 
   static contextTypes = {
     abbrNumber: PropTypes.func,
@@ -144,14 +91,6 @@ export default class SimpleBarChart extends Component {
     }
 
     return numColors > colors.length ? defaultColor : colors[index];
-  }
-
-  getMarginLeft(series) {
-    const { xAxisLabel, marginLeft, width } = this.props;
-    const longestLabel = series.data.concat(xAxisLabel).reduce((acc, { label }) =>
-      Math.max(acc, `${label}`.length)
-    , 0);
-    return Math.min(Math.min(marginLeft, longestLabel * LABEL_CHAR_WIDTH), width / 2);
   }
 
   handleShowTooltip(event, content) {
@@ -204,19 +143,47 @@ export default class SimpleBarChart extends Component {
     });
   }
 
-  renderLabel({ key, nodeHeight, x, y, value, type, index, nodeCount, maxChars }) {
+  renderLabel({ key, nodeWidth, x, y, domain, value, type, index, nodeCount }) {
     if (
       (nodeCount >= 200 && index % 10 !== 0) ||
       (nodeCount < 200 && nodeCount > 40 && index % 5 !== 0)
     ) return null;
-    const labelText = abbr(String(getLabelText(key, type)), maxChars);
+    let labelText = String(getLabelText(key, type));
+    labelText = labelText.length <= 16 ?
+      labelText : `${labelText.substring(0, 13)}…`;
 
-    const labelY = y + (nodeHeight / 2);
-    const labelX = x + (value >= 0 ? -10 : 10);
+    const labelX = x + (nodeWidth / 2);
+    if (domain[0] < 0) {
+      const labelY = value < 0 ? y - 10 : y + 10;
+      return (
+        <Text
+          textAnchor={value < 0 ? 'end' : 'start'}
+          transform={[
+            { type: 'rotate', value: [90, labelX, labelY] },
+            { type: 'translate', value: [labelX, labelY] },
+          ]}
+          {...labelFont}
+          fontWeight={get(this.state, 'hoveredNode') === key ? 700 : 400}
+          onMouseEnter={(event) => {
+            this.handleMouseEnterNode({ key, value }, event);
+          }}
+          onMouseMove={(event) => {
+            this.handleMouseEnterNode({ key, value }, event);
+          }}
+          onMouseLeave={() => {
+            this.handleMouseLeaveNode({ key });
+          }}
+        >
+          {labelText}
+        </Text>
+      );
+    }
+    const labelY = y + 10;
     return (
       <Text
-        textAnchor={value >= 0 ? 'end' : 'start'}
+        textAnchor="start"
         transform={[
+          { type: 'rotate', value: [45, labelX, labelY] },
           { type: 'translate', value: [labelX, labelY] },
         ]}
         {...labelFont}
@@ -236,20 +203,21 @@ export default class SimpleBarChart extends Component {
     );
   }
 
-  renderValueLabel({ key, nodeHeight, x, y, value, color, width }) {
+  renderValueLabel({ key, nodeWidth, x, y, value, color, height }) {
     const { valueLabelsVisible } = this.props;
     if (!valueLabelsVisible) return null;
     const labelText = heuristicRound(value);
+    const labelX = x + (nodeWidth / 2);
     const OFFSET = 5;
-    if (!labelFitsWidth(labelText, width)) return null;
+    const labelY = value < 0 ? y - OFFSET : y + OFFSET;
+    if (!labelFitsWidth(labelText, height)) return null;
 
-    const labelY = y + (nodeHeight / 2);
-    const labelX = x + (value >= 0 ? -OFFSET : OFFSET);
     return (
       <Text
+        textAnchor={value < 0 ? 'end' : 'start'}
         alignmentBaseline="center"
-        textAnchor={value >= 0 ? 'end' : 'start'}
         transform={[
+          { type: 'rotate', value: [90, labelX, labelY] },
           { type: 'translate', value: [labelX, labelY] },
         ]}
         {...labelFont}
@@ -278,9 +246,11 @@ export default class SimpleBarChart extends Component {
       onChangeVisualisationSpec,
       marginTop,
       marginRight,
+      marginLeft,
       marginBottom,
       style,
       legendVisible,
+      legendPosition,
       edit,
       padding,
       yAxisLabel,
@@ -296,13 +266,11 @@ export default class SimpleBarChart extends Component {
 
     if (!series) return null;
 
-    const marginLeft = this.getMarginLeft(series);
     const dataType = series.metadata.type;
     const paddingBottom = getPaddingBottom(series.data, dataType);
+    const dataCount = series.data.length;
     const axisLabelFontSize =
       getLabelFontSize(yAxisLabel, xAxisLabel, MAX_FONT_SIZE, MIN_FONT_SIZE, height, width);
-    const maxLabelChars = Math.floor(marginLeft / LABEL_CHAR_WIDTH);
-    const labelSizeToAxisLabelSize = Math.ceil(axisLabelFontSize / labelFont.fontSize);
 
     return (
       <ChartLayout
@@ -310,6 +278,7 @@ export default class SimpleBarChart extends Component {
         width={width}
         height={height}
         legendVisible={legendVisible}
+        legendPosition={legendPosition}
         onClick={() => {
           this.setState({ isPickingColor: undefined });
         }}
@@ -342,22 +311,19 @@ export default class SimpleBarChart extends Component {
               bottom: marginBottom,
               left: marginLeft,
             }, dimensions);
-
-            const domain = extent(series.data, ({ value }) => value);
-
-            if (domain[0] > 0) domain[0] = 0;
-            if (domain[0] < 0) margins.left = margins.right = 100;
-
             const availableHeight = dimensions.height - margins.bottom - margins.top - paddingBottom; // eslint-disable-line
             const availableWidth = dimensions.width - margins.left - margins.right;
 
-            const widthScale = scaleLinear()
+            const domain = extent(series.data, ({ value }) => value);
+            if (domain[0] > 0) domain[0] = 0;
+
+            const heightScale = scaleLinear()
               .domain([0, domain[1] - domain[0]])
-              .range([availableWidth, 0]);
+              .range([availableHeight, 0]);
 
-            const origin = widthScale(Math.abs(domain[1]));
+            const origin = heightScale(Math.abs(domain[0]));
 
-            const axisScale = scaleLinear().domain(domain).range([0, availableWidth]);
+            const axisScale = scaleLinear().domain(domain).range([0, availableHeight].reverse());
 
             const tickFormat = (value) => {
               const cutoff = 10000;
@@ -385,7 +351,7 @@ export default class SimpleBarChart extends Component {
                 <Svg width={dimensions.width} height={dimensions.height}>
 
                   {grid && (
-                    <GridColumns
+                    <GridRows
                       scale={axisScale}
                       width={availableWidth}
                       height={availableHeight}
@@ -402,17 +368,18 @@ export default class SimpleBarChart extends Component {
                       dimensions.width - margins.left - margins.right,
                       dimensions.height - margins.top - margins.bottom - paddingBottom,
                     ]}
-                    cols={1}
+                    rows={1}
                   >{nodes => (
                     <Group
                       transform={{
                         translate: [margins.left, margins.top],
                       }}
                     >
-                      {nodes.map(({ nodeHeight, y, key, value }, nodeIndex) => {
-                        const color = this.getColor(key, nodeIndex, nodes.length);
-                        const normalizedWidth = availableWidth - widthScale(Math.abs(value));
-                        const x = (value < 0) ? origin - normalizedWidth : origin;
+                      {nodes.map(({ nodeWidth, x, key, value }, i) => {
+                        const color = this.getColor(key, i, nodes.length);
+                        const normalizedHeight = availableHeight - heightScale(Math.abs(value));
+                        const colorpickerPlacement = i < dataCount / 2 ? 'right' : 'left';
+                        const y = (value < 0) ? origin : origin - normalizedHeight;
                         return (
                           <Group key={key}>
                             {(this.state.isPickingColor === key) && (
@@ -420,9 +387,13 @@ export default class SimpleBarChart extends Component {
                                 <ColorPicker
                                   title={`Pick color: ${key}`}
                                   color={color}
-                                  left={x + (normalizedWidth / 2) + margins.left}
-                                  top={y + (nodeHeight * (nodeIndex < nodes.length / 2 ? 2 : 1))}
-                                  placement={nodeIndex < nodes.length / 2 ? 'bottom' : 'top'}
+                                  left={
+                                    colorpickerPlacement === 'right' ?
+                                      margins.left + x + nodeWidth :
+                                      margins.left + x
+                                  }
+                                  top={y + (normalizedHeight / 2) + margins.top}
+                                  placement={colorpickerPlacement}
                                   onChange={({ hex }) => {
                                     onChangeVisualisationSpec({
                                       colors: { ...colorMapping, [this.state.isPickingColor]: hex },
@@ -434,13 +405,12 @@ export default class SimpleBarChart extends Component {
                             )}
                             <Rect
                               key={key}
-                              y={y + (nodeHeight * padding)}
-                              x={x}
-                              height={nodeHeight - (nodeHeight * padding * 2)}
-                              width={normalizedWidth}
+                              x={x + (nodeWidth * padding)}
+                              y={y}
+                              width={nodeWidth - (nodeWidth * padding * 2)}
+                              height={normalizedHeight}
                               fill={color}
-                              stroke="white"
-                              strokeWidth={0.1}
+                              stroke={color}
                               cursor={edit ? 'pointer' : 'default'}
                               onClick={(event) => {
                                 this.handleClickNode({ key }, event);
@@ -457,24 +427,22 @@ export default class SimpleBarChart extends Component {
                             />
                             {this.renderLabel({
                               nodeCount: series.data.length,
-                              index: nodeIndex,
+                              index: i,
                               key,
                               value,
-                              nodeHeight,
-                              y,
+                              nodeWidth,
+                              x,
+                              y: origin,
                               domain,
-                              margins,
-                              x: origin,
-                              maxChars: maxLabelChars,
                             })}
                             {this.renderValueLabel({
                               key,
                               value,
-                              nodeHeight,
-                              y,
+                              nodeWidth,
+                              x,
+                              y: origin + ((value < 0 ? 1 : -1) * normalizedHeight),
                               color,
-                              width: normalizedWidth,
-                              x: value < 0 ? x : x + normalizedWidth,
+                              height: normalizedHeight,
                             })}
                           </Group>
                         );
@@ -482,10 +450,10 @@ export default class SimpleBarChart extends Component {
                     </Group>
                   )}</Grid>
 
-                  <AxisBottom
+                  <AxisLeft
                     scale={axisScale}
                     left={margins.left}
-                    top={availableHeight + margins.top}
+                    top={margins.top}
                     label={yAxisLabel || ''}
                     stroke={'#1b1a1e'}
                     tickTextFill={'#1b1a1e'}
@@ -494,25 +462,20 @@ export default class SimpleBarChart extends Component {
                       fontSize: axisLabelFontSize,
                       textAnchor: 'middle',
                     }}
+                    labelOffset={44}
                     tickFormat={tickFormat}
                   />
 
                   <Text
                     transform={[
-                      {
-                        type: 'translate',
-                        value: [
-                          domain[0] < 0 ? origin + margins.left : margins.left - 10,
-                          margins.top - 10,
-                        ],
-                      },
+                      { type: 'translate', value: [Math.floor(this.props.width / 2), this.props.height - 10] },
                     ]}
-                    textAnchor={domain[0] < 0 ? 'middle' : 'end'}
+                    textAnchor="middle"
                     {...labelFont}
                     fontSize={axisLabelFontSize}
                     fontWeight={400}
                   >
-                    {abbr(xAxisLabel, maxLabelChars * labelSizeToAxisLabelSize)}
+                    {xAxisLabel || ''}
                   </Text>
 
                 </Svg>
