+++ client/src/components/charts/bar/StackedBarChartHorizontal.jsx
@@ -5,13 +5,15 @@ import { Rect, Svg, Group, Text } from '@potion/element';
 import get from 'lodash/get';
 import { scaleLinear } from 'd3-scale';
 import { extent } from 'd3-array';
-import { AxisLeft } from '@vx/axis';
+import { AxisBottom } from '@vx/axis';
 import { Portal } from 'react-portal';
 import merge from 'lodash/merge';
 import { stack } from 'd3-shape';
-import { GridRows } from '@vx/grid';
+import { GridColumns } from '@vx/grid';
 import itsSet from 'its-set';
+import { stackedBarPropTypes, stackedBarDefaultsProps } from './CommonBarChart';
 
+import { abbr } from '../../../utilities/utils';
 import {
   heuristicRound,
   replaceLabelIfValueEmpty,
@@ -19,15 +21,15 @@ import {
   getLabelFontSize,
   labelFitsWidth,
   labelFitsHeight,
-} from '../../utilities/chart';
-import Legend from './Legend';
-import ResponsiveWrapper from '../common/ResponsiveWrapper';
-import ColorPicker from '../common/ColorPicker';
-import ChartLayout from './ChartLayout';
-import Tooltip from './Tooltip';
-import { labelFont, MAX_FONT_SIZE, MIN_FONT_SIZE } from '../../constants/chart';
-import { isLight } from '../../utilities/color';
-import RenderComplete from './RenderComplete';
+} from '../../../utilities/chart';
+import Legend from '../Legend';
+import ResponsiveWrapper from '../../common/ResponsiveWrapper';
+import ColorPicker from '../../common/ColorPicker';
+import ChartLayout from '../ChartLayout';
+import Tooltip from '../Tooltip';
+import { labelFont, MAX_FONT_SIZE, MIN_FONT_SIZE, LABEL_CHAR_WIDTH } from '../../../constants/chart';
+import RenderComplete from '../RenderComplete';
+import { isLight } from '../../../utilities/color';
 
 const getPaddingBottom = (data) => {
   const labelCutoffLength = 16;
@@ -45,53 +47,12 @@ const getPaddingBottom = (data) => {
 
 export default class StackedBarChart extends Component {
 
-  static propTypes = {
-    data: PropTypes.shape({
-      series: PropTypes.array,
-      common: PropTypes.object,
-      metadata: PropTypes.object,
-    }),
-    colors: PropTypes.array.isRequired,
-    colorMapping: PropTypes.object,
-    onChangeVisualisationSpec: PropTypes.func,
-    width: PropTypes.number.isRequired,
-    height: PropTypes.number.isRequired,
-    legendPosition: PropTypes.oneOf(['top', 'right', 'bottom', 'left', undefined]),
-    print: PropTypes.bool,
-    interactive: PropTypes.bool,
-    edit: PropTypes.bool,
-    padding: PropTypes.number,
-    marginLeft: PropTypes.number,
-    marginRight: PropTypes.number,
-    marginTop: PropTypes.number,
-    marginBottom: PropTypes.number,
-    style: PropTypes.object,
-    legendVisible: PropTypes.bool,
-    labelsVisible: PropTypes.bool,
-    valueLabelsVisible: PropTypes.bool,
-    legendTitle: PropTypes.string,
-    yAxisLabel: PropTypes.string,
-    xAxisLabel: PropTypes.string,
-    subBucketMethod: PropTypes.string,
-    grid: PropTypes.bool,
-    yAxisTicks: PropTypes.number,
-    visualisation: PropTypes.object,
-  }
+  static propTypes = stackedBarPropTypes;
 
-  static defaultProps = {
-    interactive: true,
-    marginLeft: 70,
-    marginRight: 70,
-    marginTop: 20,
-    marginBottom: 60,
-    legendVisible: true,
-    labelsVisible: true,
-    edit: false,
-    padding: 0.1,
-    colorMapping: {},
-    grouped: false,
-    grid: true,
-  }
+  static defaultProps =
+  { ...stackedBarDefaultsProps,
+    xAxisLabel: '',
+  };
 
   static contextTypes = {
     abbrNumber: PropTypes.func,
@@ -104,7 +65,6 @@ export default class StackedBarChart extends Component {
       data: this.getData(props),
       hasRendered: false,
     };
-    this.handleMouseLeaveNode = this.handleMouseLeaveNode.bind(this);
   }
 
   componentDidMount() {
@@ -121,7 +81,7 @@ export default class StackedBarChart extends Component {
     if (!get(data, 'series[0]')) return false;
     const values = data.series[0].data
       .filter(itsSet)
-      .reduce((acc, datum, i) =>
+      .reduce((acc, { value }, i) =>
         [
           ...acc,
           data.series.reduce((acc2, series) => (itsSet(series) ? {
@@ -149,6 +109,14 @@ export default class StackedBarChart extends Component {
   getColor(key, index) {
     const { colorMapping, colors } = this.props;
     return colorMapping[key] || colors[index];
+  }
+
+  getMarginLeft(series) {
+    const { xAxisLabel, marginLeft, width } = this.props;
+    const longestLabel = series.data.concat(xAxisLabel).reduce((acc, { label }) =>
+      Math.max(acc, `${label}`.length)
+    , 0);
+    return Math.min(Math.min(marginLeft, longestLabel * LABEL_CHAR_WIDTH), width / 2);
   }
 
   handleShowTooltip(event, content) {
@@ -181,7 +149,7 @@ export default class StackedBarChart extends Component {
     const percentage = Math.round((node.values[seriesKey] / nodeTotal) * 10000) / 100;
     this.handleShowTooltip(event, [
       { key: seriesKey, color: this.getColor(seriesKey, seriesIndex), value: valueKey },
-      { key: yAxisLabel || 'y', value: `${heuristicRound(node.values[seriesKey])} (${percentage}%)` },
+      { key: yAxisLabel || 'x', value: `${heuristicRound(node.values[seriesKey])} (${percentage}%)` },
     ]);
     this.setState({ hoveredNode: { seriesKey, valueKey } });
   }
@@ -213,22 +181,19 @@ export default class StackedBarChart extends Component {
     });
   }
 
-  renderLabel({ nodeWidth, x, y, node, index, nodeCount }) {
+  renderLabel({ nodeHeight, x, y, node, index, nodeCount, maxChars }) {
     if (
       (nodeCount >= 200 && index % 10 !== 0) ||
       (nodeCount < 200 && nodeCount > 40 && index % 5 !== 0)
     ) return null;
-    let labelText = String(replaceLabelIfValueEmpty(node.key));
-    labelText = labelText.length <= 16 ?
-      labelText : `${labelText.substring(0, 13)}…`;
+    const labelText = abbr(String(replaceLabelIfValueEmpty(node.key)), maxChars);
 
-    const labelX = x + (nodeWidth / 2);
-    const labelY = y + 10;
+    const labelX = x - 10;
+    const labelY = y + (nodeHeight / 2);
     return (
       <Text
-        textAnchor="start"
+        textAnchor="end"
         transform={[
-          { type: 'rotate', value: [45, labelX, labelY] },
           { type: 'translate', value: [labelX, labelY] },
         ]}
         {...labelFont}
@@ -263,17 +228,17 @@ export default class StackedBarChart extends Component {
       labelText = heuristicRound(node.values[seriesKey]);
     }
 
-    const labelX = x + (barWidth / 2);
-    const labelY = y;
+    const labelX = x;
+    const labelY = y + (barHeight / 2);
 
-    if (!labelFitsWidth(labelText, barHeight) || !labelFitsHeight(barWidth)) return null;
+    if (!labelFitsWidth(labelText, barWidth) || !labelFitsHeight(barHeight)) return null;
 
     return (
       <Text
         textAnchor="middle"
         alignmentBaseline="center"
         transform={[
-          { type: 'rotate', value: [90, labelX, labelY] },
+          // { type: 'rotate', value: [90, labelX, labelY] },
           { type: 'translate', value: [labelX, labelY] },
         ]}
         {...labelFont}
@@ -309,9 +274,9 @@ export default class StackedBarChart extends Component {
       marginTop,
       marginRight,
       marginBottom,
-      marginLeft,
       style,
       legendVisible,
+      legendPosition,
       labelsVisible,
       legendTitle,
       edit,
@@ -321,7 +286,6 @@ export default class StackedBarChart extends Component {
       xAxisLabel,
       grid,
       visualisation,
-      legendPosition,
     } = this.props;
 
     const { tooltipItems, tooltipVisible, tooltipPosition, hasRendered } = this.state;
@@ -330,12 +294,21 @@ export default class StackedBarChart extends Component {
 
     if (!series) return null;
 
+    const marginLeft = this.getMarginLeft(series);
     const stackNodes = series.stack;
     const dataCount = series.data.length;
     const seriesCount = this.props.data.series.length;
     const paddingBottom = getPaddingBottom(series.data);
-    const axisLabelFontSize =
-      getLabelFontSize(yAxisLabel, xAxisLabel, MAX_FONT_SIZE, MIN_FONT_SIZE, height, width);
+    const axisLabelFontSize = getLabelFontSize(
+      yAxisLabel,
+      xAxisLabel,
+      MAX_FONT_SIZE,
+      MIN_FONT_SIZE,
+      height,
+      width
+    );
+    const maxLabelChars = Math.floor(marginLeft / LABEL_CHAR_WIDTH);
+    const labelSizeToAxisLabelSize = Math.ceil(axisLabelFontSize / labelFont.fontSize);
 
     return (
       <ChartLayout
@@ -358,6 +331,7 @@ export default class StackedBarChart extends Component {
                 [replaceLabelIfValueEmpty(key)]: this.getColor(key, i),
               }), {})
             }
+            // activeItem={get(this.state, 'hoveredNode.seriesKey')}
             onClick={({ datum }) => (event) => {
               this.handleClickNode(datum, event);
             }}
@@ -378,6 +352,7 @@ export default class StackedBarChart extends Component {
               bottom: marginBottom,
               left: marginLeft,
             }, dimensions);
+
             const availableHeight =
               dimensions.height - margins.bottom - margins.top - paddingBottom;
             const availableWidth = dimensions.width - margins.left - margins.right;
@@ -388,15 +363,15 @@ export default class StackedBarChart extends Component {
 
             if (domain[0] > 0) domain[0] = 0;
 
-            const heightScale = scaleLinear()
+            const widthScale = scaleLinear()
               .domain([0, domain[1]])
-              .range([availableHeight, 0]);
+              .range([0, availableWidth]);
 
-            const origin = heightScale(Math.abs(0));
+            const origin = widthScale(0);
 
             const axisScale = scaleLinear()
               .domain([0, domain[1]])
-              .range([0, availableHeight].reverse());
+              .range([0, availableWidth].reverse());
 
             const tickFormat = (value) => {
               const cutoff = 10000;
@@ -424,7 +399,7 @@ export default class StackedBarChart extends Component {
                 <Svg width={dimensions.width} height={dimensions.height}>
 
                   {grid && (
-                    <GridRows
+                    <GridColumns
                       scale={axisScale}
                       width={availableWidth}
                       height={availableHeight}
@@ -441,7 +416,7 @@ export default class StackedBarChart extends Component {
                       dimensions.width - margins.left - margins.right,
                       dimensions.height - margins.top - margins.bottom - paddingBottom,
                     ]}
-                    rows={1}
+                    cols={1}
                   >{nodes => (
                     <Group
                       transform={{
@@ -459,15 +434,13 @@ export default class StackedBarChart extends Component {
                           <Group key={seriesKey}>
                             {stackSeries.map(([y0, y1], valueIndex) => {
                               const node = nodes[valueIndex];
-                              const { nodeWidth, x, key } = node;
+                              const { nodeHeight, x, y, key } = nodes[valueIndex];
                               const color = this.getColor(seriesKey, seriesIndex);
-                              const normalizedY = heightScale(y0);
-                              const normalizedHeight = availableHeight - heightScale(y1 - y0);
+                              const normalizedX = widthScale(y0);
+                              const normalizedWidth = widthScale(y1) - normalizedX;
                               const colorpickerPlacement = valueIndex < dataCount / 2 ? 'right' : 'left';
-                              const barWidth = (nodeWidth - (nodeWidth * padding * 2)) /
+                              const barHeight = (nodeHeight - (nodeHeight * padding * 2)) /
                                 seriesCount;
-                              const normalizedX = x + (nodeWidth * padding) +
-                                (seriesIndex * barWidth);
 
                               return (
                                 <Group key={key}>
@@ -482,10 +455,10 @@ export default class StackedBarChart extends Component {
                                         color={color}
                                         left={
                                           colorpickerPlacement === 'right' ?
-                                            margins.left + x + nodeWidth :
+                                            margins.left + x + nodeHeight :
                                             margins.left + x
                                         }
-                                        top={normalizedY}
+                                        top={normalizedX}
                                         placement={colorpickerPlacement}
                                         onChange={({ hex }) => {
                                           onChangeVisualisationSpec({
@@ -501,12 +474,17 @@ export default class StackedBarChart extends Component {
                                   )}
                                   <Rect
                                     key={key}
-                                    x={normalizedX}
-                                    y={origin - normalizedHeight}
-                                    width={barWidth}
-                                    height={normalizedHeight}
+                                    y={
+                                      y +
+                                      (nodeHeight * padding) +
+                                      (seriesIndex * barHeight)
+                                    }
+                                    x={origin}
+                                    height={barHeight}
+                                    width={normalizedWidth}
                                     fill={color}
-                                    stroke={color}
+                                    stroke="white"
+                                    strokeWidth={0.1}
                                     opacity={seriesIsNotHovered ? 0.1 : 1}
                                     cursor={edit ? 'pointer' : 'default'}
                                     onClick={(event) => {
@@ -530,16 +508,18 @@ export default class StackedBarChart extends Component {
                                         event
                                       );
                                     }}
-                                    onMouseLeave={this.handleMouseLeaveNode}
+                                    onMouseLeave={() => {
+                                      this.handleMouseLeaveNode();
+                                    }}
                                   />
                                   {this.renderValueLabel({
                                     key,
-                                    barWidth,
-                                    x: normalizedX,
-                                    y: (origin - normalizedHeight) + (normalizedHeight / 2),
+                                    barWidth: normalizedWidth,
+                                    x: origin + (normalizedWidth / 2),
+                                    y: y + (nodeHeight * padding) + (seriesIndex * barHeight),
                                     value: nodes[valueIndex],
                                     color,
-                                    barHeight: normalizedHeight,
+                                    barHeight,
                                     node,
                                     seriesKey,
                                     seriesIndex,
@@ -552,31 +532,33 @@ export default class StackedBarChart extends Component {
                       })}
 
                       {nodes.map((node, index) => {
-                        const { nodeWidth, x, key } = node;
+                        const { nodeHeight, key } = node;
 
                         return (
                           <Group key={key}>
                             {this.renderLabel({
                               nodeCount: series.data.length,
                               index,
-                              nodeWidth,
-                              x,
-                              y: origin,
+                              nodeHeight,
+                              x: origin,
+                              y: nodeHeight * index,
                               domain,
                               height: 100,
                               node,
                               labelsVisible,
+                              maxChars: maxLabelChars,
                             })}
                           </Group>
                         );
                       })}
+
                     </Group>
                   )}</Grid>
 
-                  <AxisLeft
+                  <AxisBottom
                     scale={axisScale}
                     left={margins.left}
-                    top={margins.top}
+                    top={availableHeight + margins.top}
                     label={yAxisLabel || ''}
                     stroke={'#1b1a1e'}
                     tickTextFill={'#1b1a1e'}
@@ -585,19 +567,24 @@ export default class StackedBarChart extends Component {
                       fontSize: axisLabelFontSize,
                       textAnchor: 'middle',
                     }}
-                    labelOffset={44}
                     tickFormat={tickFormat}
                   />
 
                   <Text
                     transform={[
-                      { type: 'translate', value: [Math.floor(this.props.width / 2) - 125, this.props.height - 10] },
+                      {
+                        type: 'translate',
+                        value: [
+                          margins.left,
+                          margins.top - 10,
+                        ],
+                      },
                     ]}
+                    {...labelFont}
                     fontSize={axisLabelFontSize}
-                    textAnchor="middle"
-                    fontFamily="Arial"
+                    fontWeight={400}
                   >
-                    {xAxisLabel || ''}
+                    {abbr(xAxisLabel, maxLabelChars * labelSizeToAxisLabelSize)}
                   </Text>
 
                 </Svg>
@@ -618,7 +605,6 @@ export default class StackedBarChart extends Component {
       marginTop,
       marginRight,
       marginBottom,
-      marginLeft,
       style,
       legendVisible,
       legendPosition,
@@ -639,11 +625,20 @@ export default class StackedBarChart extends Component {
 
     if (!series) return null;
 
+    const marginLeft = this.getMarginLeft(series);
     const stackNodes = series.stack;
     const dataCount = series.data.length;
     const paddingBottom = getPaddingBottom(series.data);
-    const axisLabelFontSize =
-      getLabelFontSize(yAxisLabel, xAxisLabel, MAX_FONT_SIZE, MIN_FONT_SIZE, height, width);
+    const axisLabelFontSize = getLabelFontSize(
+      yAxisLabel,
+      xAxisLabel,
+      MAX_FONT_SIZE,
+      MIN_FONT_SIZE,
+      height,
+      width
+    );
+    const maxLabelChars = Math.floor(marginLeft / LABEL_CHAR_WIDTH);
+    const labelSizeToAxisLabelSize = Math.ceil(axisLabelFontSize / labelFont.fontSize);
 
     return (
       <ChartLayout
@@ -666,6 +661,7 @@ export default class StackedBarChart extends Component {
                 [replaceLabelIfValueEmpty(key)]: this.getColor(key, i),
               }), {})
             }
+            // activeItem={get(this.state, 'hoveredNode.seriesKey')}
             onClick={({ datum }) => (event) => {
               this.handleClickNode(datum, event);
             }}
@@ -686,6 +682,7 @@ export default class StackedBarChart extends Component {
               bottom: marginBottom,
               left: marginLeft,
             }, dimensions);
+
             const availableHeight =
               dimensions.height - margins.bottom - margins.top - paddingBottom;
             const availableWidth = dimensions.width - margins.left - margins.right;
@@ -696,15 +693,15 @@ export default class StackedBarChart extends Component {
 
             if (domain[0] > 0) domain[0] = 0;
 
-            const heightScale = scaleLinear()
+            const widthScale = scaleLinear()
               .domain([0, domain[1]])
-              .range([availableHeight, 0]);
+              .range([0, availableWidth]);
 
-            const origin = heightScale(Math.abs(0));
+            const origin = widthScale(0);
 
             const axisScale = scaleLinear()
               .domain([0, domain[1]])
-              .range([0, availableHeight].reverse());
+              .range([0, availableWidth]);
 
             const tickFormat = (value) => {
               const cutoff = 10000;
@@ -732,7 +729,7 @@ export default class StackedBarChart extends Component {
                 <Svg width={dimensions.width} height={dimensions.height}>
 
                   {grid && (
-                    <GridRows
+                    <GridColumns
                       scale={axisScale}
                       width={availableWidth}
                       height={availableHeight}
@@ -749,7 +746,7 @@ export default class StackedBarChart extends Component {
                       dimensions.width - margins.left - margins.right,
                       dimensions.height - margins.top - margins.bottom - paddingBottom,
                     ]}
-                    rows={1}
+                    cols={1}
                   >{nodes => (
                     <Group
                       transform={{
@@ -767,13 +764,12 @@ export default class StackedBarChart extends Component {
                           <Group key={seriesKey}>
                             {stackSeries.map(([y0, y1], valueIndex) => {
                               const node = nodes[valueIndex];
-                              const { nodeWidth, x, key } = node;
+                              const { nodeHeight, x, y, key } = nodes[valueIndex];
                               const color = this.getColor(seriesKey, seriesIndex);
-                              const normalizedY = heightScale(y0);
-                              const normalizedHeight = availableHeight - heightScale(y1 - y0);
+                              const normalizedX = widthScale(y0);
+                              const normalizedWidth = widthScale(y1) - normalizedX;
                               const colorpickerPlacement = valueIndex < dataCount / 2 ? 'right' : 'left';
-                              const barWidth = nodeWidth - (nodeWidth * padding * 2);
-                              const normalizedX = x + (nodeWidth * padding);
+                              const barHeight = (nodeHeight - (nodeHeight * padding * 2));
 
                               return (
                                 <Group key={key}>
@@ -788,10 +784,10 @@ export default class StackedBarChart extends Component {
                                         color={color}
                                         left={
                                           colorpickerPlacement === 'right' ?
-                                            margins.left + x + nodeWidth :
+                                            margins.left + x + nodeHeight :
                                             margins.left + x
                                         }
-                                        top={normalizedY}
+                                        top={normalizedX}
                                         placement={colorpickerPlacement}
                                         onChange={({ hex }) => {
                                           onChangeVisualisationSpec({
@@ -807,12 +803,13 @@ export default class StackedBarChart extends Component {
                                   )}
                                   <Rect
                                     key={key}
+                                    y={y + (nodeHeight * padding)}
                                     x={normalizedX}
-                                    y={normalizedY - normalizedHeight}
-                                    width={barWidth}
-                                    height={normalizedHeight}
+                                    height={barHeight}
+                                    width={normalizedWidth}
                                     fill={color}
-                                    stroke={color}
+                                    stroke="white"
+                                    strokeWidth={0.1}
                                     opacity={seriesIsNotHovered ? 0.1 : 1}
                                     cursor={edit ? 'pointer' : 'default'}
                                     onClick={(event) => {
@@ -836,16 +833,18 @@ export default class StackedBarChart extends Component {
                                         event
                                       );
                                     }}
-                                    onMouseLeave={this.handleMouseLeaveNode}
+                                    onMouseLeave={() => {
+                                      this.handleMouseLeaveNode();
+                                    }}
                                   />
                                   {this.renderValueLabel({
                                     key,
-                                    barWidth,
-                                    x: normalizedX,
-                                    y: normalizedY - (normalizedHeight / 2),
+                                    barWidth: normalizedWidth,
+                                    x: normalizedX + (normalizedWidth / 2),
+                                    y: y + (nodeHeight * padding),
                                     value: nodes[valueIndex],
                                     color,
-                                    barHeight: normalizedHeight,
+                                    barHeight,
                                     node,
                                     seriesKey,
                                     seriesIndex,
@@ -858,31 +857,33 @@ export default class StackedBarChart extends Component {
                       })}
 
                       {nodes.map((node, index) => {
-                        const { nodeWidth, x, key } = node;
+                        const { nodeHeight, key } = node;
 
                         return (
                           <Group key={key}>
                             {this.renderLabel({
                               nodeCount: series.data.length,
                               index,
-                              nodeWidth,
-                              x,
-                              y: origin,
+                              nodeHeight,
+                              x: origin,
+                              y: nodeHeight * index,
                               domain,
                               height: 100,
                               node,
                               labelsVisible,
+                              maxChars: maxLabelChars,
                             })}
                           </Group>
                         );
                       })}
+
                     </Group>
                   )}</Grid>
 
-                  <AxisLeft
+                  <AxisBottom
                     scale={axisScale}
                     left={margins.left}
-                    top={margins.top}
+                    top={availableHeight + margins.top}
                     label={yAxisLabel || ''}
                     stroke={'#1b1a1e'}
                     tickTextFill={'#1b1a1e'}
@@ -891,19 +892,24 @@ export default class StackedBarChart extends Component {
                       fontSize: axisLabelFontSize,
                       textAnchor: 'middle',
                     }}
-                    labelOffset={44}
                     tickFormat={tickFormat}
                   />
 
                   <Text
                     transform={[
-                      { type: 'translate', value: [Math.floor(this.props.width / 2) - 125, this.props.height - 10] },
+                      {
+                        type: 'translate',
+                        value: [
+                          margins.left,
+                          margins.top - 10,
+                        ],
+                      },
                     ]}
+                    {...labelFont}
                     fontSize={axisLabelFontSize}
-                    textAnchor="middle"
-                    fontFamily="Arial"
+                    fontWeight={400}
                   >
-                    {xAxisLabel || ''}
+                    {abbr(xAxisLabel, maxLabelChars * labelSizeToAxisLabelSize)}
                   </Text>
 
                 </Svg>
@@ -924,7 +930,6 @@ export default class StackedBarChart extends Component {
       marginTop,
       marginRight,
       marginBottom,
-      marginLeft,
       style,
       legendVisible,
       legendPosition,
@@ -945,11 +950,20 @@ export default class StackedBarChart extends Component {
 
     if (!series) return null;
 
+    const marginLeft = this.getMarginLeft(series);
     const stackNodes = series.stack;
     const dataCount = series.data.length;
     const paddingBottom = getPaddingBottom(series.data);
-    const axisLabelFontSize =
-      getLabelFontSize(yAxisLabel, xAxisLabel, MAX_FONT_SIZE, MIN_FONT_SIZE, height, width);
+    const axisLabelFontSize = getLabelFontSize(
+      yAxisLabel,
+      xAxisLabel,
+      MAX_FONT_SIZE,
+      MIN_FONT_SIZE,
+      height,
+      width
+    );
+    const maxLabelChars = Math.floor(marginLeft / LABEL_CHAR_WIDTH);
+    const labelSizeToAxisLabelSize = Math.ceil(axisLabelFontSize / labelFont.fontSize);
 
     return (
       <ChartLayout
@@ -965,13 +979,14 @@ export default class StackedBarChart extends Component {
           <Legend
             horizontal={!horizontal}
             title={legendTitle}
-            data={stackNodes.map(({ key }) => key)}
+            data={stackNodes.map(({ key }) => replaceLabelIfValueEmpty(key))}
             colorMapping={
               stackNodes.reduce((acc, { key }, i) => ({
                 ...acc,
-                [key]: this.getColor(key, i),
+                [replaceLabelIfValueEmpty(key)]: this.getColor(key, i),
               }), {})
             }
+            // activeItem={get(this.state, 'hoveredNode.seriesKey')}
             onClick={({ datum }) => (event) => {
               this.handleClickNode(datum, event);
             }}
@@ -992,24 +1007,26 @@ export default class StackedBarChart extends Component {
               bottom: marginBottom,
               left: marginLeft,
             }, dimensions);
+
             const availableHeight =
               dimensions.height - margins.bottom - margins.top - paddingBottom;
             const availableWidth = dimensions.width - margins.left - margins.right;
+
             const domain = extent(series.data, ({ values }) =>
               Object.keys(values).reduce((acc, key) => acc + Math.abs(values[key]), 0)
             );
 
             if (domain[0] > 0) domain[0] = 0;
 
-            const heightScale = scaleLinear()
+            const widthScale = scaleLinear()
               .domain([0, domain[1]])
-              .range([availableHeight, 0]);
+              .range([0, availableWidth]);
 
-            const origin = heightScale(Math.abs(0));
+            const origin = widthScale(0);
 
             const axisScale = scaleLinear()
               .domain([0, 100])
-              .range([0, availableHeight].reverse());
+              .range([0, availableWidth]);
 
             const tickFormat = value => `${value}%`;
 
@@ -1031,7 +1048,7 @@ export default class StackedBarChart extends Component {
                 <Svg width={dimensions.width} height={dimensions.height}>
 
                   {grid && (
-                    <GridRows
+                    <GridColumns
                       scale={axisScale}
                       width={availableWidth}
                       height={availableHeight}
@@ -1048,7 +1065,7 @@ export default class StackedBarChart extends Component {
                       dimensions.width - margins.left - margins.right,
                       dimensions.height - margins.top - margins.bottom - paddingBottom,
                     ]}
-                    rows={1}
+                    cols={1}
                   >{nodes => (
                     <Group
                       transform={{
@@ -1066,23 +1083,20 @@ export default class StackedBarChart extends Component {
                           <Group key={seriesKey}>
                             {stackSeries.map(([y0, y1], valueIndex) => {
                               const node = nodes[valueIndex];
-                              const { nodeWidth, x, key } = node;
-                              const individualHeightScale = scaleLinear()
-                                .range([availableHeight, 0])
+                              const { nodeHeight, x, y, key } = node;
+                              const color = this.getColor(seriesKey, seriesIndex);
+                              const individualWidthScale = scaleLinear()
+                                .range([0, availableWidth])
                                 .domain([
                                   0,
                                   Object.keys(node.values).reduce((acc, subkey) =>
                                     acc + node.values[subkey]
                                   , 0),
                                 ]);
-
-                              const color = this.getColor(seriesKey, seriesIndex);
-                              const normalizedY = individualHeightScale(y0);
-                              const normalizedHeight = availableHeight -
-                                individualHeightScale(y1 - y0);
+                              const normalizedX = individualWidthScale(y0);
+                              const normalizedWidth = individualWidthScale(y1) - normalizedX;
                               const colorpickerPlacement = valueIndex < dataCount / 2 ? 'right' : 'left';
-                              const barWidth = (nodeWidth - (nodeWidth * padding * 2));
-                              const normalizedX = x + (nodeWidth * padding);
+                              const barHeight = (nodeHeight - (nodeHeight * padding * 2));
 
                               return (
                                 <Group key={key}>
@@ -1097,10 +1111,10 @@ export default class StackedBarChart extends Component {
                                         color={color}
                                         left={
                                           colorpickerPlacement === 'right' ?
-                                            margins.left + x + nodeWidth :
+                                            margins.left + x + nodeHeight :
                                             margins.left + x
                                         }
-                                        top={normalizedY}
+                                        top={normalizedX}
                                         placement={colorpickerPlacement}
                                         onChange={({ hex }) => {
                                           onChangeVisualisationSpec({
@@ -1116,12 +1130,13 @@ export default class StackedBarChart extends Component {
                                   )}
                                   <Rect
                                     key={key}
+                                    y={y + (nodeHeight * padding)}
                                     x={normalizedX}
-                                    y={normalizedY - normalizedHeight}
-                                    width={barWidth}
-                                    height={normalizedHeight}
+                                    height={barHeight}
+                                    width={normalizedWidth}
                                     fill={color}
-                                    stroke={color}
+                                    stroke="white"
+                                    strokeWidth={0.1}
                                     opacity={seriesIsNotHovered ? 0.1 : 1}
                                     cursor={edit ? 'pointer' : 'default'}
                                     onClick={(event) => {
@@ -1145,17 +1160,18 @@ export default class StackedBarChart extends Component {
                                         event
                                       );
                                     }}
-                                    onMouseLeave={this.handleMouseLeaveNode}
+                                    onMouseLeave={() => {
+                                      this.handleMouseLeaveNode();
+                                    }}
                                   />
                                   {this.renderValueLabel({
                                     key,
-                                    barWidth,
-                                    x: normalizedX,
-                                    y: (normalizedY - normalizedHeight) +
-                                      (normalizedHeight / 2),
+                                    barWidth: normalizedWidth,
+                                    x: normalizedX + (normalizedWidth / 2),
+                                    y: y + (nodeHeight * padding),
                                     value: nodes[valueIndex],
                                     color,
-                                    barHeight: normalizedHeight,
+                                    barHeight,
                                     node,
                                     seriesKey,
                                     seriesIndex,
@@ -1168,31 +1184,33 @@ export default class StackedBarChart extends Component {
                       })}
 
                       {nodes.map((node, index) => {
-                        const { nodeWidth, x, key } = node;
+                        const { nodeHeight, key } = node;
 
                         return (
                           <Group key={key}>
                             {this.renderLabel({
                               nodeCount: series.data.length,
                               index,
-                              nodeWidth,
-                              x,
-                              y: origin,
+                              nodeHeight,
+                              x: origin,
+                              y: nodeHeight * index,
                               domain,
                               height: 100,
                               node,
                               labelsVisible,
+                              maxChars: maxLabelChars,
                             })}
                           </Group>
                         );
                       })}
+
                     </Group>
                   )}</Grid>
 
-                  <AxisLeft
+                  <AxisBottom
                     scale={axisScale}
                     left={margins.left}
-                    top={margins.top}
+                    top={availableHeight + margins.top}
                     label={yAxisLabel || ''}
                     stroke={'#1b1a1e'}
                     tickTextFill={'#1b1a1e'}
@@ -1201,19 +1219,24 @@ export default class StackedBarChart extends Component {
                       fontSize: axisLabelFontSize,
                       textAnchor: 'middle',
                     }}
-                    labelOffset={44}
                     tickFormat={tickFormat}
                   />
 
                   <Text
                     transform={[
-                      { type: 'translate', value: [Math.floor(this.props.width / 2) - 125, this.props.height - 10] },
+                      {
+                        type: 'translate',
+                        value: [
+                          margins.left,
+                          margins.top - 10,
+                        ],
+                      },
                     ]}
+                    {...labelFont}
                     fontSize={axisLabelFontSize}
-                    textAnchor="middle"
-                    fontFamily="Arial"
+                    fontWeight={400}
                   >
-                    {xAxisLabel || ''}
+                    {abbr(xAxisLabel, maxLabelChars * labelSizeToAxisLabelSize)}
                   </Text>
 
                 </Svg>
